Add epochs to the signatures account
If updating account with a different epoch, the account is first cleared
from all the data.  The idea is that client can easily reuse the account
without the need to explicitly clear it and at the same time is can send
multiple Updates in parallel.  So long as all of those Updates have the
same epoch, they won’t reset each other’s work but the first Update that
gets executed will clear up the account for all the rest of Updates.