Rename SignatureHash to SigHash
SigHash is more ‘opaque’ name which is actually better in this case.
The hash is hash of not just the signature (which wouldn’t be useful)
so a name that less strongly suggests that is better.  SigHash implies
it has something to do with signature but makse it suggestive what.

SigHash is also shorter which is additional advantage.