Return two-element array from UpdateIter
Since Message constructor expects slice of instructions, it’s more
convenient if UpdateIter::Item is an array of instructions rather than
a tuple.  If it’s a tuple, the array needs to be first constructed to
pass its address to the Message constructor.  And on the flip side,
fixed-size array can be deconstructed just like tuples can.